create separate checksum pkg
